Rename MockPackageConstraint to MockPackageContainer.Constraint
Remove MockLoadingError type

Rename MockDependencyResolver.swift to MockPackageContainer.swift
